move tooltip out of footer folder
