+++ Open/layouts/default.vue
@@ -229,7 +229,7 @@ import { BaseTeam, Team } from "../../Interfaces/team";
 import DevBanner from "../../Assets/components/DevBanner.vue";
 import TheHeader from "../../Assets/components/header/TheHeader.vue";
 import TheFooter from "../../Assets/components/footer/TheFooter.vue";
-import Tooltip from "../../Assets/components/footer/Tooltip.vue";
+import Tooltip from "../../Assets/components/Tooltip.vue";
 
 const openModule = namespace("open");
